moved gomock ItemsEq matcher to common_test and fixed comment
